[dsa] Trim ethtool stats and add port debug count
In order to reduce the load on the MDIO bus while grabbing statistics, the
list of statistics grabbed for ethtool reports is trimmed significantly.

We also add the Topaz debug_counter stat register, which is actually two
8-bit fields that track TX collisions and TX successes. This is one of two
modes this counter can operate in, so we configure it in the Global Control 2
register.

Note that the debug_counter output has to be split manually after reading
from the ethtool -S output.